Create HDFS Provisions Based on Request (#27)
Project requests will include an HDFS quota section.
POST /project
{
  "name": "Sesame",
  "purpose": "Market analysis",
  "compliance": {
    "phi": true,
    "pci": true,
    "pii": true
  },
  "hdfs": {
    "requested_gb": "20"
  }
}
GET /project/:id (incomplete)
{
  "name": "Sesame",
  "purpose": "Market analysis",
  "compliance": {
    "phi": true,
    "pci": true,
    "pii": true
  },
  "hdfs": {
    "requested_gb": 20
  },
  "created": "2017-08-01T12:23:43Z",
  "created_by": "x209812"
}
GET /project/:id (complete)
{
  "name": "Sesame",
  "purpose": "Market analysis",
  "compliance": {
    "phi": true,
    "pci": true,
    "pii": true
  },
  "hdfs": {
    "requested_gb": 20,
    "actual_gb": 20,
    "location": "/projects/sesame"
  },
  "created": "2017-08-01T12:18:43Z",
  "created_by": "x901218"
}
After the project is stored in the database, the requested amount should be formally requested.
First, the HDFS directory should be created directly below the root which is configured in the application. The name of the directory should be lowercase and all non-alpha characters should be replaced with an underscore. The location should then be added to the project record in the database.
After successful directory creation, the quota should be enforced:
$ hadoop dfsadmin -setSpaceQuota <max_size> <directory>
After the quota is acknowledged, the project record in the database should be updated.
The project table should have all three attributes recorded: requested size, actual size, and location.
The API should respond immediately acknowledging the request. The UI should poll until the projects attributes are filled in.