Fix null HiveAllocation database name
HEIM-275 (pull request #87)

When a user workspace was created and the database name in the Hive query was null.

HiveAllocation database names are empty because they are created from a join in the DB. Instead of trying to build a valid workspace from the case class, this builds it from a database query which is more likely to be correct. This is the same behavior as simple and structured workspaces.


Former-commit-id: c291a3eef4854522ff2a3cbe7f2373c77ac79c7a